+++ src/api/UserManagement.ts
@@ -366,7 +366,6 @@ export default class UserManagement extends Common {
                 result.returnValue = true;
             })
             .catch((error: AxiosError) => {
-                console.log(error);
                 if (error.response) {
                     result.statusCode = error.response.status;
                     result.error = error.response.data as ValidationError;