+++ src/views/shared/EditCreateCourseForm.vue
@@ -7,12 +7,12 @@
     <div v-else class="w-full lg:mt-20 mt-8 bg-gray-300 mx-auto h-screen">
         <button class="flex items-center mb-4 navigation-link" @click="back">
             <i class="fas text-xl fa-chevron-left"></i>
-            <span class="font-bold text-sm ml-1">Course List</span>
+            <span class="font-bold text-sm ml-1">Back</span>
         </button>
 
         <h1 class="text-2xl font-medium text-gray-700 mb-8">{{ heading }}</h1>
-
         <div>
+            <lecturer-section v-if="isAdmin" v-model:lecturerId="course.lecturerId" :error-bag="errorBag" />
             <basics-section
                 v-model:name="course.courseName"
                 v-model:type="course.courseType"
@@ -92,10 +92,11 @@
     import { CourseType } from "@/entities/CourseType";
     import { Language } from "@/entities/Language";
     import CourseManagement from "@/api/CourseManagement";
-    import { ref, computed, reactive, onBeforeMount } from "vue";
+    import { ref, reactive, computed, onBeforeMount, nextTick } from "vue";
     import DeleteCourseModal from "@/components/modals/DeleteCourseModal.vue";
     import ErrorBag from "@/use/helpers/ErrorBag";
     import ValidationResponseHandler from "@/use/helpers/ValidationResponseHandler";
+    import AccountValidationResponseHandler from "@/use/helpers/AccountValidationResponseHandler";
     import GenericResponseHandler from "@/use/helpers/GenericResponseHandler";
     import BasicsSection from "@/components/course/edit/sections/BasicsSection.vue";
     import RestrictionsSection from "@/components/course/edit/sections/RestrictionsSection.vue";
@@ -105,29 +106,19 @@
     import { Role } from "@/entities/Role";
     import UnsavedChangesModal from "@/components/modals/UnsavedChangesModal.vue";
     import { onBeforeRouteLeave } from "vue-router";
+    import LecturerSection from "@/components/course/edit/sections/LecturerSection.vue";
     import scrollToTopError from "@/use/helpers/TopError";
 
     export default {
         name: "LecturerCreateCourseForm",
         components: {
+            LecturerSection,
             BasicsSection,
             RestrictionsSection,
             TimeSection,
             DeleteCourseModal,
             UnsavedChangesModal,
             LoadingComponent,
-        },
-        async beforeRouteEnter(_to: any, _from: any, next: any) {
-            const response = await checkPrivilege(Role.LECTURER);
-
-            if (response.allowed) {
-                return next();
-            }
-            if (!response.authenticated) {
-                return next("/login");
-            }
-
-            return next("/redirect");
         },
         props: {
             editMode: {
@@ -139,6 +130,7 @@
 
         setup(props: any, { emit }: any) {
             let busy = ref(false);
+            let isAdmin = ref(false);
             let course = ref(new CourseEntity());
             let initialCourseState = new CourseEntity();
             let heading = props.editMode ? "Edit Course" : "Create Course";
@@ -178,11 +170,16 @@
             });
 
             onBeforeMount(() => {
-                getLecturerUsername();
+                askAdminRole();
                 if (props.editMode) {
                     getCourse();
                 }
             });
+
+            async function askAdminRole() {
+                const store = useStore();
+                isAdmin.value = (await store.getters.role) == Role.ADMIN;
+            }
 
             async function getLecturerUsername() {
                 const store = useStore();
@@ -225,8 +222,11 @@
             });
 
             async function createCourse() {
+                if (!isAdmin.value) {
+                    getLecturerUsername();
+                }
                 const response = await courseManagement.createCourse(course.value);
-                const handler = new ValidationResponseHandler();
+                const handler = new AccountValidationResponseHandler();
                 success.value = handler.handleReponse(response);
                 emit("update:success", success.value);
 
@@ -287,6 +287,7 @@
 
             return {
                 busy,
+                isAdmin,
                 course,
                 initialCourseState,
                 heading,