+++ tests/e2e/helpers/CourseHelper.ts
@@ -1,7 +1,13 @@
 import Course from "@/api/api_models/course_management/Course";
 import { Account } from "@/entities/Account";
+import { clear } from "console";
 import { loginAsUser } from "./AuthHelper";
-import { navigateToCourseForm, navigateToMyCoursesLecturer } from "./NavigationHelper";
+import {
+    navigateToCourseFormAdmin,
+    navigateToCourseForm,
+    navigateToMyCoursesLecturer,
+    navigateToCourseListAdmin,
+} from "./NavigationHelper";
 
 export function createCourse(course: Course) {
     navigateToCourseForm();
@@ -14,6 +20,23 @@ export function createCourse(course: Course) {
     cy.get('button[id="createCourse"]').click();
     navigateToMyCoursesLecturer();
     cy.url().should("contain", "course-management");
+    cy.wait(3000);
+    cy.get("button[title='Refresh']").click();
+    cy.get("div").contains(course.courseName).should("exist");
+}
+
+export function createCourseAdmin(course: Course) {
+    navigateToCourseFormAdmin();
+    cy.get('input[id="lecturerId"]').clear().type(course.lecturerId);
+    cy.get("input[type='radio']").get(`input[value='${course.courseType}']`).click();
+    cy.get('input[id="courseName"]').type(course.courseName);
+    cy.get("select").select(course.courseLanguage);
+    cy.get('textarea[id="courseDescription"]').type(course.courseDescription);
+    cy.get('input[id="maxParticipants"]').clear().type(course.maxParticipants.toString());
+    cy.wait(100);
+    cy.get('button[id="createCourse"]').click();
+    navigateToCourseListAdmin();
+    cy.url().should("contain", "all-courses");
     cy.wait(3000);
     cy.get("button[title='Refresh']").click();
     cy.get("div").contains(course.courseName).should("exist");
@@ -43,5 +66,20 @@ export function deleteCourse(course: Course) {
     cy.get("div").contains("Are you sure you want to delete this course?").should("exist");
     cy.get('button[id="deleteCourseModalDelete"]').click();
     cy.url().should("contain", "/course-management");
+    cy.get("div").contains(course.courseName).should("not.exist");
+}
+
+export function deleteCourseAdmin(course: Course) {
+    navigateToCourseListAdmin();
+    cy.url().should("contain", "all-courses");
+    cy.get("div").contains(course.courseName).parent().parent().find("button[id='editCourse']").click();
+    cy.wait(100);
+    cy.get("button[id='deleteCourse']").click();
+    cy.wait(100);
+    // show modal
+    cy.get("#modal-wrapper").should("exist");
+    cy.get("div").contains("Are you sure you want to delete this course?").should("exist");
+    cy.get('button[id="deleteCourseModalDelete"]').click();
+    cy.url().should("contain", "/all-courses");
     cy.get("div").contains(course.courseName).should("not.exist");
 }